separate cli output ui
